Initial lua 5.2 and 5.3 support.
It turns out the existing loader compiles and runs fine on lua 5.3. Therefore
this commit does the groundwork to link the loader against 5.3. I.e. the
basic changes to the build system to check for lua 5.3. The loader itself
is remains completely unchanged.

The logic is now: at configure time, pick the highest available lua
runtime that also provides a suitable lgi module.

This was tested on Arch Linux which ships LGI only for lua 5.3. Test suite
executes fine and I also made an example plugin.

Two option questions:
- Do we want to build multiple loaders, one per lua version?
- Do we want to rename the loader?